# Reduce the number of function calls in the `Dict` class