@timvandermeij 2019-09-24
Looks good, and also a lot more readable, to me. Thanks!